Implement BloodMagic orbs as recipe Ingredients.
While OrbRegistry is marked as deprecated, it is the main method used to
access ItemStacks for each orb above a certain tier. Otherwise using
orbs in recipes requires chaining .or statements together with a
pre-defined list of orbs above the tier that you wish to use.

Implements a bracket handler for <orb:master> and
<orb:bloodmagic:master>. <orb:master> is automatically converted into
<orb:bloodmagic:master>.

Implements mods.bloodmagic.Orbs.getOrb("name") which returns an Ingredient
equivalent to using the bracket <orb:name>.

Probably not the best nor correct way to implement this, but it
functions.